+++ src/index.ts
@@ -1,9 +1,11 @@
 import axios from 'axios';
+import { AgentContentMessage, AgentStatusMessage, Suggestion } from './message';
 import { verifyMiddleware } from './verify';
 
 export interface DireqtApiConfiguration {
     signingSecret: string;
     accessToken: string;
+    _messagingApiRoot?: string;
 }
 
 export class DireqtApi {
@@ -18,11 +20,24 @@ export class DireqtApi {
 }
 
 export class DireqtMessagingApi {
-    constructor(private config: DireqtApiConfiguration) {}
+    private readonly apiRoot: string = `https://gateway.direqt.io`;
+    private readonly path = `/agent/direqt/receiver`;
 
-    public async sendTextMessage(userId: string, text: string) {
-        const url = `https://gateway.direqt.io/agent/direqt/receiver`;
+    constructor(private config: DireqtApiConfiguration) {
+        if (config._messagingApiRoot) {
+            this.apiRoot = config._messagingApiRoot;
+        }
+    }
 
+    private get url() {
+        return `${this.apiRoot}${this.path}`;
+    }
+
+    public async sendTextMessage(
+        userId: string,
+        text: string,
+        suggestions?: Suggestion[]
+    ) {
         const body = {
             userId,
             agentMessage: {
@@ -30,7 +45,30 @@ export class DireqtMessagingApi {
                 content: {
                     contentType: 'text',
                     text,
-                },
+                } as AgentContentMessage,
+            },
+        };
+
+        if (suggestions) {
+            body.agentMessage.content.suggestions = suggestions;
+        }
+
+        const params = {
+            access_token: this.config.accessToken,
+        };
+
+        return await axios.post(this.url, body, { params });
+    }
+
+    public async sendContentMessage(
+        userId: string,
+        contentMessage: AgentContentMessage
+    ) {
+        const body = {
+            userId,
+            agentMessage: {
+                messageType: 'content',
+                content: contentMessage,
             },
         };
 
@@ -38,7 +76,26 @@ export class DireqtMessagingApi {
             access_token: this.config.accessToken,
         };
 
-        return await axios.post(url, body, { params });
+        return await axios.post(this.url, body, { params });
+    }
+
+    public async sendStatusMessage(
+        userId: string,
+        statusMessage: AgentStatusMessage
+    ) {
+        const body = {
+            userId,
+            agentMessage: {
+                messageType: 'status',
+                status: statusMessage,
+            },
+        };
+
+        const params = {
+            access_token: this.config.accessToken,
+        };
+
+        return await axios.post(this.url, body, { params });
     }
 
     public verifyMiddleware() {