+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Direqt SDK for JavaScript - Change Log
 
+## v1.2.0
+ * Allow suggestions to be sent with messaging.sendTextMessage
+ * Add sendContentMessage and sendStatusMessage helper methods to messaging interface.
+ * Add richmedia sample
+
+## v1.0.1
+ * First public release
+
 ## v0.3.0
  * Updated to use messaging API